Infobars (3/4): Implement prioritization and queuing logic
This change implements the core logic for infobar prioritization within
the `InfoBarContainer`, controlled by the `kInfobarPrioritization`
feature flag. This system prevents high-priority infobars, such as those
related to security, from being queued behind less critical
notifications.

When enabled, a new queuing mechanism is activated with the following
rules:
- Critical: Can stack concurrently up to a configurable limit. These
  are always prioritized.
- Default: Single-visible. Additional infobars are queued and will
  not be shown while any critical infobars are visible or pending.
- Low: Single-visible. These are queued behind any pending critical or
  default infobars.

When a visible infobar is closed, the container automatically promotes
the next highest-priority infobar from the queue. A comprehensive unit
test suite is added to verify this new behavior.

Additionally, this patch refactors the feature parameter handling and
applies a const-correctness fix to the `EqualsDelegate` method.

BYPASS_LARGE_CHANGE_WARNING: This change need to be together so revert can be easier.

Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/7097862
Commit-Queue: Daniel Soromou <koretadaniel@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1546385}

+++ components/infobars/core/features.cc
@@ -8,24 +8,24 @@ namespace infobars {
 
 BASE_FEATURE(kInfobarPrioritization, base::FEATURE_DISABLED_BY_DEFAULT);
 
-BASE_FEATURE_PARAM(int,
+BASE_FEATURE_PARAM(size_t,
                    kMaxVisibleCritical,
                    &kInfobarPrioritization,
                    "max_visible_critical",
                    2);
 
-BASE_FEATURE_PARAM(int,
+BASE_FEATURE_PARAM(size_t,
                    kMaxVisibleDefault,
                    &kInfobarPrioritization,
                    "max_visible_default",
                    1);
 
-BASE_FEATURE_PARAM(int,
+BASE_FEATURE_PARAM(size_t,
                    kMaxVisibleLow,
                    &kInfobarPrioritization,
                    "max_visible_low",
                    1);
-BASE_FEATURE_PARAM(int,
+BASE_FEATURE_PARAM(size_t,
                    kMaxLowQueued,
                    &kInfobarPrioritization,
                    "max_low_queued",
